chore: adding module deprecation warnings; aliasing applicationspec to arc32contract
