+++ src/algokit_utils/clients/client_manager.py
@@ -15,8 +15,9 @@ from algokit_utils._legacy_v2.application_specification import ApplicationSpecif
 from algokit_utils.applications.app_client import AppClient
 from algokit_utils.applications.app_deployer import AppLookup
 from algokit_utils.applications.app_factory import AppFactory, AppFactoryParams
+from algokit_utils.applications.app_spec.arc56 import Arc56Contract
 from algokit_utils.clients.dispenser_api_client import TestNetDispenserApiClient
-from algokit_utils.models.application import AppClientParams, Arc56Contract
+from algokit_utils.models.application import AppClientParams
 from algokit_utils.models.network import AlgoClientConfig, AlgoClientConfigs
 from algokit_utils.models.state import TealTemplateParams
 from algokit_utils.protocols.client import AlgorandClientProtocol
