+++ src/algokit_utils/errors/__init__.py
@@ -1,0 +1,5 @@
+from algokit_utils.errors.logic_error import LogicError
+
+__all__ = [
+    "LogicError",
+]
